Rename project lookup service files to match new naming scheme
