+++ src/extension-host/extension-host.ts
@@ -9,7 +9,7 @@ import dataProviderService from '@shared/services/data-provider.service';
 import extensionAssetService from '@shared/services/extension-asset.service';
 import { getErrorMessage } from '@shared/utils/util';
 import { CommandNames } from 'papi-shared-types';
-import { startProjectLookupService } from '@extension-host/services/project-lookup-backend.service';
+import { startProjectLookupService } from '@extension-host/services/project-lookup.service.host';
 
 // #region Test logs
