+++ src/renderer/services/papi-frontend.service.ts
@@ -12,7 +12,7 @@ import { papiNetworkService, PapiNetworkService } from '@shared/services/network
 import { papiWebViewService, PapiWebViewService } from '@shared/services/web-view.service';
 import internetService, { InternetService } from '@shared/services/internet.service';
 import dataProviderService, { DataProviderService } from '@shared/services/data-provider.service';
-import { ProjectLookupServiceType } from '@shared/models/project-lookup.model';
+import { ProjectLookupServiceType } from '@shared/services/project-lookup.service.model';
 import projectLookupService from '@shared/services/project-lookup.service';
 import {
   papiFrontendProjectDataProviderService,
